services/s2-consumer/requirements.txt
kafka-python==2.0.2
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
pymongo==4.4.0
cassandra-driver==3.25.0
